requirements.txt
Django~=3.2.12
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
drf-yasg~=1.21.7
django-filter~=23.5
python-dotenv~=1.0.0
psycopg2-binary~=2.9.9
gunicorn~=21.2.0
